# openapi.admin.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/customers": {
      "get": {
        "operationId": "AdminController_listCustomers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page (50 rows per page).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match on name, externalRef and contactEmails. Applied before pagination; `total` reflects the filtered count.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerListResponseModel"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "google-sso": []
          },
          {
            "admin-token": []
          }
        ],
        "summary": "List customers (paginated, 50/page, sorted by name/externalRef).",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Roles are validated against the audiences; a role change takes effect on the next publish/rollout only (no auto-rollout here). Optional `acceptedVersions` records already signed versions as IMPORT acceptances. externalRef is unique only among customers with OVERLAPPING roles — the partner and customer external ID spaces are separate, so the same externalRef may coexist on records with disjoint roles; a duplicate that shares at least one role → 422 INVALID_STATE.",
        "operationId": "AdminController_createCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerBodyModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCustomerResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Body validation failed (strict schema).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "VERSION_NOT_FOUND (acceptedVersions)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "INVALID_STATE (externalRef duplicate sharing a role, invalid e-mail) · UNKNOWN_AUDIENCE · ROLE_MISMATCH",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "google-sso": []
          },
          {
            "admin-token": []
          }
        ],
        "summary": "Create a customer",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customers/{id}": {
      "get": {
        "description": "The full customer record (id, externalRef, firstName, lastName, companyName, roles, contactEmails) — e.g. for the detail-page header.",
        "operationId": "AdminController_getCustomer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerRowModel"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "CUSTOMER_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "google-sso": []
          },
          {
            "admin-token": []
          }
        ],
        "summary": "Get a single customer by id",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "description": "A `roles` change takes effect on the next publish/rollout only (no auto-rollout on role add). Adding a role is rejected with 422 INVALID_STATE if it would overlap another customer that already shares this externalRef.",
        "operationId": "AdminController_updateCustomer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerBodyModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerRowModel"
                }
              }
            }
          },
          "400": {
            "description": "Body validation failed (strict schema).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "CUSTOMER_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "INVALID_STATE (invalid e-mail, externalRef overlap on role add) · UNKNOWN_AUDIENCE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "google-sso": []
          },
          {
            "admin-token": []
          }
        ],
        "summary": "Update a customer (subset of firstName/lastName/companyName/roles/contactEmails)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customers/{id}/acceptance-links": {
      "post": {
        "description": "Returns `${PUBLIC_BASE_URL}/accept/<token>` — a capability URL the admin sends directly to the person who has to accept (no portal integration needed). The raw token appears only in this response; the server persists just its SHA-256. Optional `audienceKey` scopes the page to one audience; `expiresInDays` defaults to 30 (max 365). Writes an ACCEPTANCE_LINK_CREATE audit entry.",
        "operationId": "AdminController_createAcceptanceLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAcceptanceLinkBodyModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAcceptanceLinkResponseModel"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "CUSTOMER_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "INVALID_STATE (PUBLIC_BASE_URL unset, expiresInDays out of range) · UNKNOWN_AUDIENCE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "google-sso": []
          },
          {
            "admin-token": []
          }
        ],
        "summary": "Mint a hosted acceptance link for a customer",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "clickwrap-server — admin API",
    "description": "Administration surface (`/admin/**`): documents & versions, publish/rollout, customers, operations (overview, history, manual acceptance, deadlines/reminders) and the dynamic entities (audiences, document types). Errors are always `{ code, message }` — the codes are documented in docs/API.md §7.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "google-sso": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Google ID token (Google SSO; domain/allowlist checked server-side)."
      },
      "admin-token": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "Static admin token fallback (dev/CI; ADMIN_API_TOKEN). Optional: x-admin-user names the actor."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "INVALID_STATE",
            "description": "Stable DomainErrorCode (see docs/API.md §7)."
          },
          "message": {
            "type": "string",
            "example": "A customer with externalRef \"ext-1\" and an overlapping role already exists"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CustomerListResponseModel": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerRowModel"
            }
          },
          "total": {
            "type": "number",
            "example": 173
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
...
      "UpdateCustomerBodyModel": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Jane",
            "description": "Contact person's given name."
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Contact person's family name."
          },
          "companyName": {
            "type": "string",
            "example": "Acme GmbH",
            "description": "Optional company/organisation name."
          },
          "roles": {
            "description": "Takes effect on the next publish/rollout.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "contactEmails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}